requirements.txt
python-dotenv==1.0.1
gunicorn==23.0.0
Flask==3.0.3
gevent==24.10.3
Flask-SocketIO==5.4.1
youtube-transcript-api==0.6.2
yt-dlp==2024.10.22
openai-whisper==20240930